feat: refactor project folder structure
