+++ backend/src/main/java/org/conexaotreinamento/conexaotreinamentobackend/dto/request/PatchExerciseRequestDTO.java
@@ -1,4 +1,4 @@
-package org.conexaotreinamento.conexaotreinamentobackend.api.dto.request;
+package org.conexaotreinamento.conexaotreinamentobackend.dto.request;
 
 import jakarta.validation.constraints.Size;
